# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CloudStorage.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Files": {
      "get": {
        "tags": [
          "Files"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Files"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Files/search": {
      "get": {
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Files/{id}": {
      "get": {
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Files/{id}/download": {
      "get": {
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Files/{id}/permissions": {
      "post": {
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilePermissionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FilePermissionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FilePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Files/{id}/share": {
      "post": {
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilePermissionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FilePermissionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FilePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Files/{id}/download-link": {
      "get": {
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FilePermissionDto": {
        "required": [
          "permissionType",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "permissionType": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "FileUploadDto": {
        "required": [
          "contentType",
          "fileName",
          "size"
        ],
        "type": "object",
        "properties": {
          "fileName": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "contentType": {
            "minLength": 1,
            "type": "string"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int64"
          },
          "chunkCount": {
            "type": "integer",
            "format": "int32"
          },
          "hash": {
            "type": "string",
            "nullable": true
          },
          "folderId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}